// frontend/src/app/lib/pocketbase-types.ts
export enum Collections {
	Authorigins = "_authOrigins",
	Externalauths = "_externalAuths",
	Mfas = "_mfas",
	Otps = "_otps",
	Superusers = "_superusers",
	Carriers = "carriers",
	Partners = "partners",
	QuoteRequests = "quote_requests",
	ShippingMethods = "shipping_methods",
	SupportedCountries = "supported_countries",
	Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string
export type IsoAutoDateString = string & { readonly autodate: unique symbol }
export type RecordIdString = string
export type FileNameString = string & { readonly filename: unique symbol }
export type HTMLString = string

type ExpandType<T> = unknown extends T
	? T extends unknown
		? { expand?: unknown }
		: { expand: T }
	: { expand: T }

// System fields
export type BaseSystemFields<T = unknown> = {
	id: RecordIdString
	collectionId: string
	collectionName: Collections
} & ExpandType<T>
// ...
export type QuoteRequestsRecord = {
	completed?: boolean
	created: IsoAutoDateString
	created_by?: RecordIdString
	declared_value_usd?: number
	delivery_instructions?: string
	destination_country?: RecordIdString
	height_inches?: number
	id: string
	item_image?: FileNameString
	length_inches?: number
	quote_price?: number
	recipient_address_1?: string
	recipient_city?: string
	recipient_first_name?: string
	recipient_last_name?: string
	recipient_phone?: string
	recipient_postal_code?: string
	recipient_state?: string
	referrer?: RecordIdString
	rejected?: boolean
	rejected_reason?: string
	sender_address?: string
	sender_city?: string
	sender_email?: string
	sender_first_name?: string
	sender_last_name?: string
	sender_phone?: string
	sender_state?: string
	sender_zip_code?: string
	updated: IsoAutoDateString
	weight_lbs?: string
	width_inches?: number
}
// ...
export type QuoteRequestsResponse<Texpand = unknown> = Required<QuoteRequestsRecord> & BaseSystemFields<Texpand>